+++ services/web-app/src/js/pages/transactions/index.tsx
@@ -1,4 +1,4 @@
-import { Fragment } from 'react';
+import { Fragment, useState } from 'react';
 import useAuthenticatedUser from '../../hooks/useAuthenticatedUser';
 import LoadingState from '../../enums/loadingState';
 import LoginRedirect from '../login/redirect';
@@ -7,11 +7,13 @@ import { Alert, Box, Link } from '@mui/material';
 import TransactionsHeader from './header';
 import { extensionId } from '../../constants';
 import '../../../css/transactions.scss';
-import TransactionsList from './transactions-list';
+import TransactionsList from './list';
+import TransactionsSummary from './summary';
 
 export default function Transactions() {
   const authenticatedUser = useAuthenticatedUser();
-  const [startDate, endDate] = [new Date('1/1/2021'), new Date()];
+  const [startDate, setStartDate] = useState<Date>(new Date('1/1/2021'));
+  const [endDate, setEndDate] = useState<Date>(new Date());
 
   if (authenticatedUser.loadingState !== LoadingState.Success) {
     // Loading and failure states are handled by the parent.
@@ -45,6 +47,12 @@ export default function Transactions() {
       <CreatorTabs />
       <Box sx={{ p: 1, flexGrow: 1, flexWrap: 'wrap' }}>
         <TransactionsHeader />
+        <TransactionsSummary
+          startDate={startDate}
+          endDate={endDate}
+          setStartDate={setStartDate}
+          setEndDate={setEndDate}
+        />
         <TransactionsList startDate={startDate} endDate={endDate} />
       </Box>
     </Box>
